// src/types/types.ts
export type newsType = {
    author: string
    authors: []
    clean_url: string
    country: string
    is_opinion: boolean
    language: string
    link: string
    media: string
    published_date: string
    published_date_precision: string
    rank: number
    rights: string
    summary: string
    title: string
    topic: string
    twitter_account: string
    id: string
    score: number
}